requirements.txt
requests
PyGithub==1.51